Add exact measure after at_most measure if results are smaller than constraint
