+++ Xamarin.Forms.Platform.Android/AppCompat/Platform.cs
@@ -8,6 +8,7 @@ using Android.Views;
 using Android.Views.Animations;
 using ARelativeLayout = Android.Widget.RelativeLayout;
 using Xamarin.Forms.Internals;
+using Debug = System.Diagnostics.Debug;
 
 namespace Xamarin.Forms.Platform.Android.AppCompat
 {
@@ -169,42 +170,38 @@ namespace Xamarin.Forms.Platform.Android.AppCompat
 			Performance.Start(reference);
 
 			// FIXME: potential crash
-			IVisualElementRenderer viewRenderer = Android.Platform.GetRenderer(view);
+			IVisualElementRenderer visualElementRenderer = Android.Platform.GetRenderer(view);
 
 			// negative numbers have special meanings to android they don't to us
 			widthConstraint = widthConstraint <= -1 ? double.PositiveInfinity : _context.ToPixels(widthConstraint);
 			heightConstraint = heightConstraint <= -1 ? double.PositiveInfinity : _context.ToPixels(heightConstraint);
 
-			var wmode = MeasureSpecMode.AtMost;
-			var hmode = MeasureSpecMode.AtMost;
+			bool widthConstrained = !double.IsPositiveInfinity(widthConstraint);
+			bool heightConstrained = !double.IsPositiveInfinity(heightConstraint);
 
-			if(view is View fview)
-			{
-				if (fview.HorizontalOptions.Alignment == LayoutAlignment.Fill)
-				{
-					wmode = MeasureSpecMode.Exactly;
-				}
-				if (fview.VerticalOptions.Alignment == LayoutAlignment.Fill)
-				{
-					hmode = MeasureSpecMode.Exactly;
-				}
-			}
-
-			int width = !double.IsPositiveInfinity(widthConstraint)
-							? MeasureSpecFactory.MakeMeasureSpec((int)widthConstraint, wmode)
+			int widthMeasureSpec = widthConstrained
+							? MeasureSpecFactory.MakeMeasureSpec((int)widthConstraint, MeasureSpecMode.AtMost)
 							: MeasureSpecFactory.MakeMeasureSpec(0, MeasureSpecMode.Unspecified);
 
-			int height = !double.IsPositiveInfinity(heightConstraint)
-							 ? MeasureSpecFactory.MakeMeasureSpec((int)heightConstraint, hmode)
+			int heightMeasureSpec = heightConstrained
+							 ? MeasureSpecFactory.MakeMeasureSpec((int)heightConstraint, MeasureSpecMode.AtMost)
 							 : MeasureSpecFactory.MakeMeasureSpec(0, MeasureSpecMode.Unspecified);
 
-			SizeRequest rawResult = viewRenderer.GetDesiredSize(width, height);
+			SizeRequest rawResult = visualElementRenderer.GetDesiredSize(widthMeasureSpec, heightMeasureSpec);
 			if (rawResult.Minimum == Size.Zero)
 				rawResult.Minimum = rawResult.Request;
 			var result = new SizeRequest(new Size(_context.FromPixels(rawResult.Request.Width), _context.FromPixels(rawResult.Request.Height)),
 				new Size(_context.FromPixels(rawResult.Minimum.Width), _context.FromPixels(rawResult.Minimum.Height)));
 
+			if ((widthConstrained && result.Request.Width < widthConstraint)
+				|| (heightConstrained && result.Request.Height < heightConstraint))
+			{
+				// Do a final exact measurement in case the native control needs to fill the container
+				(visualElementRenderer as IViewRenderer)?.MeasureExactly();
+			}
+
 			Performance.Stop(reference);
+
 			return result;
 		}
