+++ Xamarin.Forms.Platform.Android/Renderers/FormsTextView.cs
@@ -28,37 +28,6 @@ namespace Xamarin.Forms.Platform.Android
 		{
 		}
 
-		//protected override void OnDraw(Canvas canvas)
-		//{
-		//	System.Diagnostics.Debug.WriteLine($">>>>> FormsTextView OnDraw: {canvas.Width}, {canvas.Height}");
-		//	base.OnDraw(canvas);
-		//}
-
-		//protected override void OnLayout(bool changed, int left, int top, int right, int bottom)
-		//{
-		//	System.Diagnostics.Debug.WriteLine($">>>>> FormsTextView OnLayout: changed is {changed}, ltrb = {left}, {top}, {right}, {bottom}");
-		//	base.OnLayout(changed, left, top, right, bottom);
-		//}
-
-		//protected override void OnMeasure(int widthMeasureSpec, int heightMeasureSpec)
-		//{
-		//	var widthMode =	MeasureSpec.GetMode(widthMeasureSpec);
-		//	var heightMode = MeasureSpec.GetMode(heightMeasureSpec);
-
-		//	var wmode = ((MeasureSpecMode)widthMode).ToString();
-		//	var hmode = ((MeasureSpecMode)heightMode).ToString();
-
-		//	var width = MeasureSpecFactory.GetSize(widthMeasureSpec);
-		//	var height = MeasureSpecFactory.GetSize(heightMeasureSpec);
-		//	System.Diagnostics.Debug.WriteLine($">>>>> FormsTextView OnMeasure: {wmode} {width}, {hmode} {height}");
-		//	base.OnMeasure(widthMeasureSpec, heightMeasureSpec);
-
-		//	var mwidth = MeasuredWidth;
-		//	var mheight = MeasuredHeight;
-
-		//	System.Diagnostics.Debug.WriteLine($">>>>> FormsTextView OnMeasure result: {mwidth}, {mheight}");
-		//}
-
 		public override void Invalidate()
 		{
 			if (!_skip)
